Encapsulated About Feature
Added two new classes to About (AboutActions and AboutFeature)
	- AboutActions: SoC Principle, this class encapsulates About actions that were on the Notepad Actions class
	- AboutFeature: This class is responsible for interacting with the Notepad istance menu and other Notepad itens, as well with the About classes that implement the About feature, to fully implement the Feature on the notepad. This class also provides a single method to be used on the Notepad class to add the About feature.

MenuBar as been turned static so that it could be used by static methods on the AboutFeature class. Both the Notepad and its Menu, as well any other itens on the Notepad are unique (no multiple menus, toolbars, or notepad instance), thus this should not be a problem.